Move CoerceSized from marker to ops
So it can be right beside CoerceUnsized
