+++ src/liballoc/rc.rs
@@ -252,10 +252,10 @@ use core::fmt;
 use core::hash::{Hash, Hasher};
 use core::intrinsics::abort;
 use core::marker;
-use core::marker::{Unsize, CoerceSized, PhantomData};
+use core::marker::{Unsize, PhantomData};
 use core::mem::{self, align_of_val, forget, size_of_val};
 use core::ops::Deref;
-use core::ops::CoerceUnsized;
+use core::ops::{CoerceUnsized, CoerceSized};
 use core::ptr::{self, NonNull};
 use core::convert::From;
 use core::usize;
